tune tf8 l6

+++ test_tfbind10_tune_23_2.py
@@ -125,7 +125,7 @@ def main():
     best_tf8_hyper = np.unique(best_tf8_hyper,axis=0)
     print(len(best_tf8_hyper))
     num_candidates = len(best_tf8_hyper)
-    best_tf8_hyper = best_tf8_hyper[2*num_candidates/5:3*num_candidates/5] 
+    best_tf8_hyper = best_tf8_hyper[int(2*num_candidates/5):int(3*num_candidates/5)] 
 
     sampling_lr = 0.05
     for lengthscale in [6.0]:
